+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.12.0] - 2018-01-23
+
 ### Added
 
 - `fx.App` wrapper so that closer does not have to be manually populated
@@ -220,7 +223,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bootstrap options
 
 
-[Unreleased]: https://github.com/goph/fxt/compare/v0.11.0...HEAD
+[Unreleased]: https://github.com/goph/fxt/compare/v0.12.0...HEAD
+[0.12.0]: https://github.com/goph/fxt/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/goph/fxt/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/goph/fxt/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/goph/fxt/compare/v0.8.0...v0.9.0